good dtls is ok but no audio
good dtls is ok but no audio

+++ webrtcgateway/glib/include/glib-2.0/gobject/gparam.h
@@ -163,8 +163,7 @@ typedef enum
   G_PARAM_STATIC_BLURB	      = 1 << 7,
   /* User defined flags go here */
   G_PARAM_EXPLICIT_NOTIFY     = 1 << 30,
-  /* Avoid warning with -Wpedantic for gcc6 */
-  G_PARAM_DEPRECATED          = (gint)(1u << 31)
+  G_PARAM_DEPRECATED          = 1 << 31
 } GParamFlags;
 
 /**
